Code simplification
- Remove StreamTrait
- Bugfix header from AbstractCsv::output to be HTTP/2 compliant
