+++ tests/CsvTest.php
@@ -69,8 +69,8 @@ class CsvTest extends TestCase
         // Due to the variety of ways the xdebug expresses Content-Type of text files,
         // we cannot count on complete string matching.
         $this->assertContains('content-type: text/csv', strtolower($headers[0]));
-        $this->assertSame($headers[1], 'Content-Transfer-Encoding: binary');
-        $this->assertSame($headers[2], 'Content-Disposition: attachment; filename="test.csv"');
+        $this->assertSame($headers[1], 'content-transfer-encoding: binary');
+        $this->assertSame($headers[2], 'content-disposition: attachment; filename="test.csv"');
     }
 
     public function testToString()
